+++ src/os/infrastructure/console_command/message_command.ts
@@ -1,20 +1,7 @@
 import { OperatingSystem } from "os/os"
+import { isMessageObserver } from "os/infrastructure/message_observer"
 import { parseProcessId } from "./command_utility"
 import { ConsoleCommand, CommandExecutionResult } from "./console_command"
-
-export interface MessageObserver {
-  /**
-   *
-   * @param message
-   * @returns 標準出力に表示する実行結果
-   */
-  didReceiveMessage(message: string): string
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/explicit-module-boundary-types
-export function isMessageObserver(arg: any): arg is MessageObserver {
-  return arg.didReceiveMessage !== undefined
-}
 
 export class MessageCommand implements ConsoleCommand {
   public constructor(
