+++ src/task/procedural.ts
@@ -1,0 +1,8 @@
+export interface Procedural {
+  runOnTick(): void
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/explicit-module-boundary-types
+export function isProcedural(arg: any): arg is Procedural {
+  return arg.runOnTick !== undefined
+}
